+++ src/Controller/CategoryController.php
@@ -9,6 +9,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 
 #[Route('/category')]
 class CategoryController extends AbstractController
@@ -21,9 +22,13 @@ class CategoryController extends AbstractController
         ]);
     }
 
-    #[Route('/super/nouveau', name: 'category_new', methods: ['GET', 'POST'])]
+    #[IsGranted("ROLE_ADMIN")]
+    #[Route('/new', name: 'category_new', methods: ['GET', 'POST'])]
     public function new(Request $request, CategoryRepository $categoryRepository): Response
     {
+
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $category = new Category();
         $form = $this->createForm(CategoryType::class, $category);
         $form->handleRequest($request);
@@ -51,9 +56,13 @@ class CategoryController extends AbstractController
         ]);
     }
 
-    #[Route('/{idcategory}/super/edit', name: 'category_edit', methods: ['GET', 'POST'])]
+    #[IsGranted("ROLE_ADMIN")]
+    #[Route('/{idcategory}/edit', name: 'category_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Category $category, CategoryRepository $categoryRepository): Response
     {
+
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $form = $this->createForm(CategoryType::class, $category);
         $form->handleRequest($request);
 
@@ -70,9 +79,13 @@ class CategoryController extends AbstractController
         ]);
     }
 
+    #[IsGranted("ROLE_ADMIN")]
     #[Route('/{idcategory}', name: 'category_delete', methods: ['POST'])]
     public function delete(Request $request, Category $category): Response
     {
+
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         if ($this->isCsrfTokenValid('delete'.$category->getIdcategory(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($category);
